Make ltotape thread-safe
The ltotape driver uses a thread-shared ltotape_scsi_io_type structure
that contains fields involved in SCSI operations: cdb, cdb_length, data,
data_length, data_direction, actual_data_length, sensedata, sense_length,
timeout_ms.

As a result the driver is not re-entrant, while libltfs expects it to
be. We fix that situation by adding a wrapper layer that copies the
ltotape_scsi_io_type structure to the stack when doing SCSI operations,
so that the offending fields are not shared among threads anymore.